LPS-55928 move generic query impl classes
